extract common object detail and table settings
